Add functionality to "archiving" older status of a file
A new table, file_history is used to store older versions of the
file, so that a backup can be restored from an arbitrary point in
time.

The functionality has been added to the update script, so that
when it runs, when it detects the status of a file has changed,
it updates it, archiving the older status (file status, path,
timestamps) in the file_history table.

This will not detect renames/loss files yet- they will be stored
as new files (without storing them twice on backup storage).

This also assumes there are no duplicate files with the same title,
sha1 and upload timestamp- which is not true at the time.

WIP: T299764

+++ mediabackups/cli/backup_wiki.py
@@ -22,22 +22,18 @@ TMPDIR = '/srv/mediabackup'
 
 class DownloadException(Exception):
     """Raised on errors while trying to download the file from production"""
-    pass
 
 
 class UploadException(Exception):
     """Raised on errors while trying to uplaod the file into the backup storage"""
-    pass
 
 
 class DuplicateException(Exception):
     """Raised in case the file already exists on the backup storage"""
-    pass
 
 
 class EncryptionException(Exception):
     """Raised in case attempting to encrypt a file fails"""
-    pass
 
 
 def download_file_from_production(f, tmp_dir):
@@ -153,7 +149,8 @@ def main():
                     'id': file_id, 'file': f, 'status': new_status, 'location': location}
                 )
                 try:
-                    os.remove(download_path)
+                    if download_path is not None:
+                        os.remove(download_path)
                 except OSError:
                     pass  # ignoring errors as the file may not exist
         metadata.update_status(status_list)
